[RC] modify bi_vgg net 2

+++ deep_id/bi_2_base.py
@@ -853,11 +853,11 @@ class NN:
         axis = list(range(len(x_shape) - 1))
 
         with tf.variable_scope('batch_normal'):
-            beta = tf.Variable(np.zeros(params_shape, dtype=np.float64), name='beta', dtype=tf.float64)
-            gamma = tf.Variable(np.ones(params_shape, dtype=np.float64), name='gamma', dtype=tf.float64)
+            beta = tf.Variable(np.zeros(params_shape, dtype=np.float32), name='beta', dtype=tf.float32)
+            gamma = tf.Variable(np.ones(params_shape, dtype=np.float32), name='gamma', dtype=tf.float32)
 
-            moving_mean = tf.Variable(np.zeros(params_shape, dtype=np.float64), name='moving_mean', trainable=False, dtype=tf.float64)
-            moving_variance = tf.Variable(np.ones(params_shape, dtype=np.float64), name='moving_variance', trainable=False, dtype=tf.float64)
+            moving_mean = tf.Variable(np.zeros(params_shape, dtype=np.float32), name='moving_mean', trainable=False, dtype=tf.float32)
+            moving_variance = tf.Variable(np.ones(params_shape, dtype=np.float32), name='moving_variance', trainable=False, dtype=tf.float32)
 
             # beta = self.get_variable('beta', params_shape, initializer=tf.zeros_initializer)
             # gamma = self.get_variable('gamma', params_shape, initializer=tf.ones_initializer)
